🚸 Remove star button in Bookmarks
Rationale: removing an item from the bookmarks is not often used.
Removing this button cleans up space and prevent accidental click error
when trying to add an item to the cart.
Adding/removing an item to the bookmarks is still possible from the item
page, it will only require one more click.