🐛 Fix wrong data submitted
This is a weird issue with react-hook-form and reset.
If you update the quantity by clicking on the arrow inside the field and
thus keeping the focus in another field, when you submit the quantity
field value would not be updated.
This does not happen if you don't use reset(). It still happens if you
use it inside useEffect like recommended in the doc.
Workaround: use latest values from the form when submitting.